get-data-lambda/requirements.txt
# For Lambda functions packaging only
asgiref==3.7.2
boto3==1.28.59
botocore==1.31.59
certifi==2023.7.22
charset-normalizer==3.3.0
idna==3.4
jmespath==1.0.1
numpy==1.26.0
packaging==23.2
pandas==2.1.1
psycopg2-binary==2.9.9
requests==2.31.0
s3transfer==0.7.0
six==1.16.0
sqlparse==0.4.4
tenacity==8.2.3
tzdata==2023.3